Make __title__ optional
In many cases, the registered PyPI name will be the same as
package/module name, and we will usually be able to construct that
from the passed path.  This commit fills in the path-based value as a
default, so users who do set __title__ explicitly will be able to
clobber that default with their custom value.